modify in logics

+++ src/main/java/com/plh/foodappbackend/model/CartItem.java
@@ -1,5 +1,6 @@
 package com.plh.foodappbackend.model;
 
+import java.math.BigDecimal;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -10,7 +11,7 @@ import lombok.NoArgsConstructor;
 public class CartItem {
     private String foodId;
     private int quantity;
-    private double price; // Price at the time of adding to cart
+    private BigDecimal price; // Price at the time of adding to cart
     private String name;
     private String imageUrl;
 }
